+++ src/sspi/pku2u/cert_utils/mod.rs
@@ -1,0 +1,6 @@
+pub mod validation;
+#[cfg(target_os = "windows")]
+pub mod win_extraction;
+
+#[cfg(target_os = "windows")]
+pub use win_extraction as extraction;
